to be common reloading widget
