+++ lib/presentation/pages/top/top_page.dart
@@ -7,6 +7,7 @@ import 'package:count_habits/presentation/pages/theme/color_schemes.dart';
 import 'package:count_habits/presentation/pages/top/components/add_new_one_card.dart';
 import 'package:count_habits/presentation/pages/top/components/animated_counter.dart';
 import 'package:count_habits/presentation/pages/top/components/summary_card.dart';
+import 'package:count_habits/util/reloading_widget.dart';
 import 'package:dots_indicator/dots_indicator.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -104,7 +105,7 @@ class TopPage extends ConsumerWidget {
           fontSize: 18,
           textColor: Colors.white,
         );
-        return const _ReLoadingWidget();
+        return ReLoadingWidget(() => ref.invalidate(countersProvider));
       },
       loading: () {
         final images = Assets.images.values;
@@ -123,34 +124,6 @@ class TopPage extends ConsumerWidget {
           ),
         );
       },
-    );
-  }
-}
-
-// TODO 共通パーツ検討
-class _ReLoadingWidget extends ConsumerWidget {
-  const _ReLoadingWidget();
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final theme = ref.watch(cupertinoThemeProvider);
-    return CupertinoPageScaffold(
-      child: Center(
-        child: CupertinoButton(
-          color: theme.barBackgroundColor,
-          onPressed: () => ref.invalidate(countersProvider),
-          borderRadius: BorderRadius.circular(10),
-          child: Text(
-            '再読み込み',
-            style: TextStyle(
-              color: theme.brightness == Brightness.light ? Colors.black : Colors.white,
-              fontSize: 15,
-              fontWeight: FontWeight.w600,
-              fontFamily: theme.textTheme.textStyle.fontFamily,
-            ),
-          ),
-        ),
-      ),
     );
   }
 }
